refactor: move lap state and calc up to main
