+++ src/components/Laps/LapsDisplay.jsx
@@ -8,25 +8,8 @@ const initialHighestLowestLaps = {
   highestLap: { id: undefined, interval: 0 },
   lowestLap: { id: undefined, interval: Infinity },
 }
-const initialRunningLap = {
-  interval: 0,
-  id: 0,
-}
-const initialEmptyLapsState = [1, 2, 3, 4, 5, 6]
 
-/* useReducer function */
-const reducerRunningLap = (state, action) => {
-  switch (action.type) {
-    case 'increment':
-      return { interval: state.interval + 1, id: state.id }
-    case 'add lap':
-      return { interval: 0, id: state.id + 1 }
-    case 'reset':
-      return { interval: 0, id: 1 }
-    default:
-      throw new Error('wrong action type')
-  }
-}
+const initialEmptyLapsState = [1, 2, 3, 4, 5, 6]
 
 const reducerHighestLowest = (state, action) => {
   switch (action.type) {
@@ -41,20 +24,9 @@ const reducerHighestLowest = (state, action) => {
   }
 }
 
-// TODO: Move lap state back to parent :(
-
 function LapControls(props) {
-  const { elapsedTime, lapId } = props
+  const { runningLap, allLaps } = props
 
-  const runningLapRef = useRef()
-  const [stateRunningLap, dispatchRunningLap] = useReducer(
-    reducerRunningLap,
-    initialRunningLap,
-  )
-
-  const [allLaps, setAllLaps] = useState([])
-
-  const highestLowestRef = useRef()
   const [stateHighestLowest, dispatchHighestLowest] = useReducer(
     reducerHighestLowest,
     initialHighestLowestLaps,
@@ -63,56 +35,25 @@ function LapControls(props) {
   const [emptyLaps, setEmptyLaps] = useState(initialEmptyLapsState)
   const [isScrolling, setIsScrolling] = useState(false)
 
+  useEffect(() => {
+    const newLap = allLaps[0]
+    if (allLaps.length) findHighestLowestLaps(newLap)
+  }, [allLaps])
+
   const findHighestLowestLaps = (newLap) => {
-    if (newLap.interval < highestLowestRef.current.lowestLap.interval) {
+    if (newLap.interval < stateHighestLowest.lowestLap.interval) {
       dispatchHighestLowest({ type: 'change lowest', lowestLap: newLap })
     }
-    if (newLap.interval > highestLowestRef.current.highestLap.interval) {
+    if (newLap.interval > stateHighestLowest.highestLap.interval) {
       dispatchHighestLowest({ type: 'change highest', highestLap: newLap })
     }
   }
 
-  /* Track refs */
-  useEffect(() => {
-    highestLowestRef.current = stateHighestLowest
-  }, [stateHighestLowest])
-
-  useEffect(() => {
-    runningLapRef.current = stateRunningLap
-  }, [stateRunningLap])
-
-  /* Keep running lap in time with elapsedTime */
-  useEffect(() => {
-    if (elapsedTime) dispatchRunningLap({ type: 'increment' })
-  }, [elapsedTime])
-
-  /* Fires every time a lap is added or timer is reset */
-  useEffect(() => {
-    switch (lapId) {
-      /* Reset */
-      case 0:
-        resetLaps()
-        break
-      /* First running lap */
-      case 1:
-        setEmptyLaps((prevArray) => prevArray.slice(0, -1))
-        break
-      default:
-        const newLap = runningLapRef.current
-        setAllLaps((prevAllLaps) => [newLap, ...prevAllLaps])
-        findHighestLowestLaps(newLap)
-
-        dispatchRunningLap({ type: 'add lap' })
-
-        setEmptyLaps((prevArray) => prevArray.slice(0, -1))
-    }
-  }, [lapId])
-
-  const resetLaps = () => {
-    dispatchRunningLap({ type: 'reset' })
-    dispatchHighestLowest({ type: 'reset' })
-    setAllLaps([])
-    setEmptyLaps(initialEmptyLapsState)
+  /* Helper function to get highest / lowest class name */
+  const getClassName = (id) => {
+    if (allLaps.length > 1 && stateHighestLowest.highestLap.id === id) return 'highest'
+    if (allLaps.length > 1 && stateHighestLowest.lowestLap.id === id) return 'lowest'
+    return ''
   }
 
   /* Workaround to fade the scrollbar */
@@ -126,23 +67,14 @@ function LapControls(props) {
     })
   }, [])
 
-  /* Helper function to get highest / lowest class name */
-  const getClassName = (id) => {
-    if (allLaps.length > 1 && highestLowestRef.current.highestLap.id === id)
-      return 'highest'
-    if (allLaps.length > 1 && highestLowestRef.current.lowestLap.id === id)
-      return 'lowest'
-    return ''
-  }
-
   return (
     <section className={`lap-container ${isScrolling ? 'scrollbar-fade' : ''}`}>
       <table className={'lap-table'}>
         <tbody id={'lap-list'}>
-          {stateRunningLap.interval > 0 && (
+          {runningLap.interval > 0 && (
             <tr className={'lap'}>
-              <td>{`Lap ${stateRunningLap.id}`}</td>
-              <td>{getFormattedTime(stateRunningLap.interval)}</td>
+              <td>{`Lap ${runningLap.id}`}</td>
+              <td>{getFormattedTime(runningLap.interval)}</td>
             </tr>
           )}
           {allLaps.map((lap) => (
@@ -155,6 +87,7 @@ function LapControls(props) {
               <td>{getFormattedTime(lap.interval)}</td>
             </tr>
           ))}
+          {/* EmptyLaps component that will have informed the number of laps by props */}
           {emptyLaps &&
             emptyLaps.map((emptyLap, index) => (
               <tr key={index} className={'lap'}>
