+++ src/components/Main/Main.jsx
@@ -1,28 +1,55 @@
 import React from 'react'
-import { useState } from 'react'
+import { useState, useEffect, useReducer } from 'react'
 
 import TimerControls from '../Buttons/TimerControls'
 import TimerDisplay from '../Timer/TimerDisplay'
 import LapsDisplay from '../Laps/LapsDisplay'
 import Footer from '../Footer/Footer'
 
+const initialRunningLap = {
+  interval: 0,
+  id: 1,
+}
+
+const reducerRunningLap = (state, action) => {
+  switch (action.type) {
+    case 'increment':
+      return { interval: state.interval + 1, id: state.id }
+    case 'add lap':
+      return { interval: 0, id: state.id + 1 }
+    case 'reset':
+      return { interval: 0, id: 1 }
+    default:
+      throw new Error('wrong action type')
+  }
+}
+
 function Main() {
   const [elapsedTime, setElapsedTime] = useState(0)
-  const [lapId, setLapId] = useState(0)
+  const [stateRunningLap, dispatchRunningLap] = useReducer(
+    reducerRunningLap,
+    initialRunningLap,
+  )
+  const [allLaps, setAllLaps] = useState([])
 
   const handleTime = (passedTime) => {
-    /* Starting lap -> make LapsDisplay update the first running lap */
-    if (passedTime === 1) setLapId(1)
     setElapsedTime(passedTime)
   }
 
+  useEffect(() => {
+    if (elapsedTime) dispatchRunningLap({ type: 'increment' })
+  }, [elapsedTime])
+
   const handleAddLap = () => {
-    setLapId((prevId) => prevId + 1)
+    const newLap = stateRunningLap
+    setAllLaps((prevAllLaps) => [newLap, ...prevAllLaps])
+    dispatchRunningLap({ type: 'add lap' })
   }
 
   const reset = () => {
     setElapsedTime(0)
-    setLapId(0)
+    setAllLaps([])
+    dispatchRunningLap({ type: 'reset' })
   }
 
   return (
@@ -33,7 +60,7 @@ function Main() {
         handleLap={handleAddLap}
         handleReset={reset}
       />
-      <LapsDisplay elapsedTime={elapsedTime} lapId={lapId} />
+      <LapsDisplay runningLap={stateRunningLap} allLaps={allLaps} />
       <Footer />
     </main>
   )
